+++ lizard_measure/management/commands/import_krw_portaal.py
@@ -19,7 +19,6 @@ from lizard_area.models import DataAdministrator
 from lizard_measure.models import KRWStatus
 from lizard_measure.models import KRWWatertype
 from lizard_measure.models import WaterBody
-from lizard_measure.models import OrganizationType
 from lizard_measure.models import Organization
 from lizard_measure.models import Unit
 from lizard_measure.models import MeasuringRod
@@ -54,7 +53,11 @@ def _records(xml_filename):
 
 def _get_or_create(model, get_kwargs, extra_kwargs={}):
     """
-    Return object, created_boolean
+    Return object, created_boolean.
+
+    TODO: Get rid of this method, since django's get_or_create
+    does just the same. You need to specify the 'extra_kwargs'
+    as a keyword argument 'defaults'...
     """
     try:
         return model.objects.get(**get_kwargs), False
@@ -111,18 +114,13 @@ def import_KRW_lookup(filename):
     """
     Import various domains into seperate lizard_measure models
     """
-    organization_type, organization_type_created = _get_or_create(
-        model=OrganizationType,
-        get_kwargs={'name': 'Geïmporteerd uit KRW-portaal'}
-    )
     for rec in _records(filename):
         # Insert 'uitvoerders'
         if rec['domein'] == 'uitvoerder':
             organization, organization_created = _get_or_create(
                 model=Organization,
-                get_kwargs={'name': rec['description']},
+                get_kwargs={'description': rec['description']},
                 extra_kwargs={
-                    'organization_type': organization_type,
                     'source': Organization.SOURCE_KRW_PORTAL,
                 }
             )
@@ -182,7 +180,7 @@ def import_measure_types(filename):
         # Add the units
         units = rec['eenheid'].split(', ')
         for u_str in units:
-            unit_obj = Unit.objects.get_or_create(unit=u_str)[0]
+            unit_obj = Unit.objects.get_or_create(code=u_str)[0]
             measure_type.units.add(unit_obj)
 
 
@@ -327,16 +325,11 @@ def import_measures(filename):
         )
 
         unit, unit_created = Unit.objects.get_or_create(
-            unit=rec['mateenh'],
+            code=rec['mateenh'],
         )
 
-#       executive, executive_created = _get_or_create(
-#           model=Organization,
-#           get_kwargs={'name': rec['uitvoerder']},
-#       )
-
         # vvv Decided from examination of screenshots from KRW portal
-        initiator = Organization.objects.get(name=rec['uitvoerder'])
+        initiator = Organization.objects.get(description=rec['uitvoerder'])
 
         datetime_in_source = datetime.datetime.strptime(
             rec['datum'],
@@ -436,7 +429,7 @@ def import_measures(filename):
                 continue
             cost_carrier = rec['kostendrager' + n]
             cost_percentage = _to_float_or_none(rec['kostenpercent' + n])
-            organization = Organization.objects.get(name=cost_carrier)
+            organization = Organization.objects.get(description=cost_carrier)
             funding_organization = FundingOrganization(
                 percentage=cost_percentage,
                 organization=organization,